Precompute colors
It turns out that SRGB computation is expensive – I think both the blending and
the float → integer conversion, but it might've just been one.  So storing an
array of the possible colors only takes a little extra space (twice as much, in
the worst case) for a small but noticeable speed improvement.  (About 3.5
seconds for 1000×1000 images, in my rough-'n'-ready testing.)